chore: extract logic to find which packages have changed into it's own module
