+++ scripts/check-next-version.js
@@ -1,48 +1,11 @@
 import core from "@actions/core"
-import {$} from "zx"
-import {promises as fs} from "fs"
+import {packagesWhichChanged} from "./lib/packagesWhichChanged"
 import {context} from "@actions/github"
 
-let patch = Symbol('patch')
-let minor = Symbol('minor')
-let major = Symbol('major')
-
 try {
-    $.verbose = false
-    const publishablePackages = await getPublishablePackages()
-    // Map of package names to version bump type
-    // E.G. 'components/o-quote' => Symbol(major)
-    const packagesToPublish = new Map;
     let baseRef = getBaseRef()
     let headRef = getHeadRef()
-    let commitHashesAndMessages = await commitHashesAndMessagesBetween(baseRef, headRef)
-    for (const commitHashAndMessage of commitHashesAndMessages) {
-        const [hash, message] = splitOnce(commitHashAndMessage, ' ')
-        if (isPublishableCommitMessage(message)) {
-            let version = parseCommitMessageToVersion(message)
-            let filesChanged = await findModifiedFilesInCommit(hash)
-            for (const pkg of publishablePackages) {
-                for (const file of filesChanged) {
-                    if (file.startsWith(pkg)) {
-                        if (packagesToPublish.has(pkg)) {
-                            const currentVersion = packagesToPublish.get(pkg);
-                            // If the version if patch, we can skip all the rest as nothing is lower than a patch
-                            if (version === patch) {
-                                continue
-                            } else if (version === minor && currentVersion === patch) {
-                                packagesToPublish.set(pkg, version);
-                                // If the version is majoe, we can assign it immediately as nothign is higher than a major
-                            } else if (version === major) {
-                                packagesToPublish.set(pkg, version);
-                            }
-                        } else {
-                            packagesToPublish.set(pkg, version);
-                        }
-                    }
-                }
-            }
-        }
-    }
+    const packagesToPublish = await packagesWhichChanged(baseRef, headRef)
     if (packagesToPublish.size) {
         core.notice('A commit is a `fix` or `feat` commit, a release will be made.')
         console.log({packagesToPublish})
@@ -51,32 +14,6 @@ try {
     }
 } catch (error) {
 	core.setFailed(error.message)
-}
-
-function splitOnce(string, separator) {
-    let [first, ...rest] = string.split(separator)
-    if (rest.length === 0) {
-        return [first]
-    } else {
-        return [first, rest.join(separator)]
-    }
-}
-
-async function getPublishablePackages() {
-    let releasePleaseFile
-    try {
-        releasePleaseFile = await fs.readFile('release-please-config.json', "utf-8")
-    } catch (error) {
-        throw new Error('cannot read release-please-config.json')
-    }
-    let releasePleaseConfig
-    try {
-        releasePleaseConfig = JSON.parse(releasePleaseFile)
-    } catch (error) {
-        throw new Error('release-please-config.json contains invalid json')
-    }
-    const publishablePackages = Object.keys(releasePleaseConfig.packages)
-    return publishablePackages
 }
 
 function getBaseRef() {
@@ -92,33 +29,4 @@ function getHeadRef() {
         headRef = `origin/${context.payload.pull_request.head.ref}`
     }
     return headRef
-}
-async function commitHashesAndMessagesBetween(baseRef, headRef) {
-    let commitHashesAndMessages = (await $`git log --pretty="format:%H %s" ${baseRef}~20...${headRef} --`).stdout.split('\n')
-    return commitHashesAndMessages
-}
-
-async function findModifiedFilesInCommit(commitSha) {
-    let filesChanged = (await $`git show --pretty="format:" --name-only ${commitSha}`).stdout.split('\n')
-    return filesChanged
-}
-
-function isPublishableCommitMessage(message) {
-    if (message.startsWith('fix:') || message.startsWith('fix!:') || message.startsWith('feat:') || message.startsWith('feat!:')) {
-        return true
-    } else {
-        return false
-    }
-}
-
-function parseCommitMessageToVersion(message) {
-    let version;
-    if (message.startsWith('fix:')){
-        version = patch
-    } else if (message.startsWith('feat:')) {
-        version = minor
-    } else {
-        version = major
-    }
-    return version
 }
